Fix issue preventing highest resolution results.
Fix an issue preventing the `Facebook` plugin from selecting the image
with the highest resolution available if a download link is present.
Note that, if no download link is available on the image page, the
displayed photo is selected instead which may or may not be the one with
the highest resolution available.