Rewrite plugin build script.
Removed redundancy by combining both the filter generation and the
plugin combination. That way the plugins only need to be read once.

New script is also Python 3 compatible, and is tested on both Python 2.7
and 3.2. Both yield the same results, apart from unimportant ordering
differences due to the `mergeex` module.

Ant build script was amended to use the new script.