Add build process to generate sandboxes.
Sandbox scripts are stored within `/plugins/sandboxes/` and are
automatically converted and merged into the plugin files. This allows an
easier editing process for built-in sandboxes as no manual decoding and
encoding process is required. All sandbox scripts can be edited directly
in JavaScript source files without worrying about the JSON structure or
its rules.

Script is Python 3 compatible and tested in Python 2.7 and Python 3.2.

Sandbox scripts are automatically assigned to a plugin `<name>.json` if
their file names are either `<name>.process.js` or `<name>.resolve.js`.
The `process` or `resolve` part decides as which plugin parameter the
script is set. It is possible to specify both for a single plugin.